UPDATE: rework zpl generate (NOT STABLE)

+++ Src/Infrastructure/Ws.Database.Nhibernate/Entities/Print/Labels/SqlLabelMap.cs
@@ -94,14 +94,14 @@ internal sealed class SqlLabelMap : ClassMapping<Label>
         Property(x => x.Kneading, m =>
         {
             m.Column("KNEADING");
-            m.Type(NHibernateUtil.Int16);
+            m.Type(NHibernateUtil.Int32);
             m.NotNullable(true);
         });
 
         Property(x => x.BundleCount, m =>
         {
             m.Column("BUNDLE_COUNT");
-            m.Type(NHibernateUtil.UInt16);
+            m.Type(NHibernateUtil.Int32);
             m.NotNullable(true);
         });
